+++ src/commands/index.ts
@@ -13,6 +13,7 @@ import { XcodeCommand } from './xcode/xcode.ts';
 import { ClickUpCommand } from './clickup.ts';
 import TestCommand from './test/index.ts';
 import YouTubeCommand from './youtube/index.ts';
+import NotionCommand from './notion/index.ts';
 
 export const commands: CommandMap = {
   web: new WebCommand(),
@@ -28,4 +29,5 @@ export const commands: CommandMap = {
   clickup: new ClickUpCommand(),
   test: new TestCommand(),
   youtube: new YouTubeCommand(),
+  notion: NotionCommand,
 };